requirements.txt
funcodec==0.2.0
nltk==3.9.1
numpy==1.24.4
openai-whisper==20231117
pip==24.2
PyYAML==6.0.1
scipy==1.10.1
soundfile==0.12.1
torch==2.0.1
torchaudio==2.0.2
torchmetrics==1.2.0
tqdm==4.66.4
modelscope==1.14.0
librosa==0.10.1
